requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-ollama>=1.0.0
python-dotenv>=1.0.0
ollama>=0.1.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
